+++ osidb/models/flaw/flaw.py
@@ -865,7 +865,7 @@ class Flaw(
         """
         check if the flaw has a workflow label with the given name
         """
-        from osidb.models.flaw.label_v2 import WorkflowLabel
+        from osidb.models.flaw.label import WorkflowLabel
 
         return WorkflowLabel.objects.filter(flaw=self, name=label).exists()
 
@@ -1075,6 +1075,7 @@ class Flaw(
             self.task_key = jtq.create_or_update_task(self)
             self.adjust_classification(save=False)
             Flaw.objects.filter(uuid=self.uuid).update(
+                auto_timestamps=False,
                 task_key=self.task_key,
                 workflow_name=self.workflow_name,
                 workflow_state=self.workflow_state,