+++ docs/CHANGELOG.md
@@ -5,9 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- migrate alias-typed manual-triage labels to workflow type (OSIDB-5335)
+- bulk add approved workflow label to the flaws resolved before Argus (OSIDB-5334)
+- make V2 labels be just labels (OSIDB-5320)
+
 ### Fixed
 - Gate ACE behind workflow eligibility to prevent spurious affects on manually-triaged,
   untriaged, and legacy flaws (OSIDB-4972, OSIDB-5307, OSIDB-5312)
+
+### Removed
+- delete deprecated alerts (OSIDB-5337)
 
 ## [5.14.0] - 2026-08-05
 ### Changed